Move group_atom_indices to TrajectoryUtils
